Lazy init of EE info
