+++ packages/core/strapi/ee/resources/key.pub
